Changed some endpoints details

+++ backend/src/main/java/com/PAP_team_21/flashcards/controllers/FriendshipController.java
@@ -13,7 +13,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -24,22 +23,28 @@ public class FriendshipController {
     private final FriendshipRepository friendshipRepository;
     private final CustomerRepository customerRepository;
 
-    @GetMapping("/{id}")
+    @GetMapping("/getFriendship/{id}")
     @JsonView(JsonViewConfig.Public.class)
-    public ResponseEntity<?> getAllFriendships(Authentication authentication, @PathVariable int id) {
+    public ResponseEntity<?> getFriendships(Authentication authentication, @PathVariable int id) {
         String email = authentication.getName();
         Optional<Customer> customerOpt = customerRepository.findByEmail(email);
         if(customerOpt.isEmpty())
         {
             return ResponseEntity.badRequest().body("No user with this id found");
         }
+        Customer customer = customerOpt.get();
 
         Optional<Friendship> friendshipOpt = friendshipRepository.findById(id);
         if (friendshipOpt.isEmpty()) {
             return ResponseEntity.badRequest().body("No friendship with this id found");
         }
+        Friendship friendship = friendshipOpt.get();
 
-        return ResponseEntity.ok(friendshipOpt.get());
+        if (friendship.getReceiverId() == customer.getId() || friendship.getSenderId() == customer.getId()) {
+            return ResponseEntity.ok(friendship);
+        }
+
+        return ResponseEntity.badRequest().body("User do not have access to this friendship");
     }
 
     @PostMapping("/create")
@@ -51,12 +56,15 @@ public class FriendshipController {
         {
             return ResponseEntity.badRequest().body("No user with this id found");
         }
+        Customer customer = customerOpt.get();
 
-        Friendship friendship = new Friendship(request.getSenderId(), request.getReceiverId(),
-                                                request.isAccepted());
+        if (request.getSenderId() == customer.getId()) {
+            Friendship friendship = new Friendship(request.getSenderId(), request.getReceiverId());
+            friendshipRepository.save(friendship);
+            return ResponseEntity.ok(friendship);
+        }
 
-        friendshipRepository.save(friendship);
-        return ResponseEntity.ok(friendship);
+        return ResponseEntity.ok("User do not have access to this friendship");
     }
 
     @PostMapping("/update")
@@ -68,6 +76,7 @@ public class FriendshipController {
         {
             return ResponseEntity.badRequest().body("No user with this id found");
         }
+        Customer customer = customerOpt.get();
 
         Optional<Friendship> friendshipOpt = friendshipRepository.findById(request.getFriendshipId());
         if (friendshipOpt.isEmpty()) {
@@ -75,10 +84,14 @@ public class FriendshipController {
         }
 
         Friendship friendship = friendshipOpt.get();
-        friendship.setAccepted(request.isAccepted());
 
-        friendshipRepository.save(friendship);
-        return ResponseEntity.ok(friendship);
+        if (friendship.getReceiverId() == customer.getId() || friendship.getSenderId() == customer.getId()) {
+            friendship.setAccepted(request.isAccepted());
+            friendshipRepository.save(friendship);
+            return ResponseEntity.ok(friendship);
+        }
+
+        return ResponseEntity.badRequest().body("User do not have access to this friendship");
     }
 
     @DeleteMapping("/delete")
@@ -89,6 +102,7 @@ public class FriendshipController {
         {
             return ResponseEntity.badRequest().body("No user with this id found");
         }
+        Customer customer = customerOpt.get();
 
         Optional<Friendship> friendshipOpt = friendshipRepository.findById(friendshipId);
         if (friendshipOpt.isEmpty()) {
@@ -96,8 +110,12 @@ public class FriendshipController {
         }
         Friendship friendship = friendshipOpt.get();
 
-        friendshipRepository.delete(friendship);
-        return ResponseEntity.ok("Friendship successfully deleted");
+        if (friendship.getReceiverId() == customer.getId() || friendship.getSenderId() == customer.getId()) {
+            friendshipRepository.delete(friendship);
+            return ResponseEntity.ok("Friendship successfully deleted");
+        }
+
+        return ResponseEntity.badRequest().body("User do not have access to this friendship");
     }
 
 }
